CodeClean: CodeReview: I-L
git-svn-id: svn://svn.gathered-geeks.org/emulator/trunk@144 06a71e81-df49-0410-adf8-df517167f62f

+++ src/IRE.cpp
@@ -2,11 +2,11 @@
 
 IRE::IRE(RawOperation *rawOp) : ConvertOperation(rawOp)
 {
-    /*nichts zu tun*/
+    /* nichts zu tun */
 }
 
 void IRE::run(Configuration *c)
 {
-    c->getAC()->setFloat( (float) c->getAC()->getInt() );
-    c->setPC(c->getPC() +1);
+    c->getAC()->setFloat(c->getAC()->getInt());
+    c->setPC(c->getPC() + 1);
 }
